backend authentication/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.28.3
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
pydantic==2.8.2
